backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10
qdrant-client==1.13.3
litellm==1.72.0
sentence-transformers==3.3.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
cryptography==44.0.0
websockets==14.1
tiktoken==0.8.0
markdown-it-py==3.0.0
aiodocker==0.22.2
aiofiles==24.1.0
PyMuPDF==1.25.1
mcp>=1.3.0
Pillow>=10.0.0
pytesseract>=0.3.13
openai-whisper>=20231117
mutagen>=1.47.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
marker-pdf>=1.0.0
celery[redis]==5.4.0
redis[hiredis]>=5.0.0
croniter>=1.3.0
gitpython>=3.1.0
PyYAML>=6.0
